Plotting done.
The dimension of the matrix and number of integers per page is
hardcoded in gen_data.sh. Change that in the file if you want a new
data set. A new directory is created
output_data_<dimension>_<integers_per_page> and data is generated
in that directory. Plot is also generated in that directory.

So to run, change the two variables in gen_data.sh and ./gen_data.sh.
Make sure you make the script executable.